% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_program_stages.R
\name{get_program_stage_table}
\alias{get_program_stage_table}
\title{get_program_stage_table}
\usage{
get_program_stage_table()
}
\value{
data.frame
}
\description{
always returns a table
}
\author{
Reinhard Simon
}

