% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_program_stages.R
\name{new_program_stage_table}
\alias{new_program_stage_table}
\title{new_program_stage_table}
\usage{
new_program_stage_table()
}
\value{
data.frame
}
\description{
creates an empty table with dummy values
}
\author{
Reinhard Simon
}

