% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_program_stages.R
\name{post_program_stage_table}
\alias{post_program_stage_table}
\title{post_program_stage_table}
\usage{
post_program_stage_table(table_program_stages)
}
\arguments{
\item{table_program_stages}{a data frame}
}
\description{
store program_stage table
}
\author{
Reinhard Simon
}

