% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server_program_stages.R
\name{server_program_stages}
\alias{server_program_stages}
\title{server_program_stages}
\usage{
server_program_stages(input, output, session, dom = "hot_programstage")
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{dom}{target dom element name}
}
\description{
Constructs table
}
\author{
Reinhard Simon
}

