% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ui_program_stage.R
\name{ui_program_stage}
\alias{ui_program_stage}
\title{ui_program_stage}
\usage{
ui_program_stage(type = "tab", title = "Program stage configuration",
  name = "resource_program_stage", output = "hot_programstage")
}
\arguments{
\item{type}{of ui Element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}

\item{output}{name of output element}
}
\description{
returns a re-usable user interface element
}
\author{
Reinhard Simon
}

